surrealdb-chatbot-rag/pyproject.toml
[project]
name = "surrealdb-chatbot-rag"
version = "0.1.0"
description = "Simple RAG chatbot with SurrealDB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "surrealdb>=1.0.8",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
